requirements.txt
# dbt==0.20.0
pandas==1.3.0
pre-commit==2.14.0
python-dotenv==0.19.0
spotipy==2.19.0
